First commit for InspectorModel and Inspector3.
Lots of changes.  I'm eliminating Inspector2 in favor of my third
Inspector class.  This one attempts to implement MVC pattern (which is
difficult on Android) and is a LOT cleaner.  Inspector3 also should be
faster as I've eliminated some redundant code and made the Inflator a
class var.

Furthermore Inspector3 no longer tries to maintain a static AsyncTask.
I realize that was an unnecessary complication.  Whenever there's a
configuration change, everything starts all over with the new Activity,
including a new AsyncTask (the old AsyncTask is warned about the config
change and will stop itself).